Move markdown in editmode
