+++ superset-frontend/cypress-base/cypress/utils/index.ts
@@ -109,3 +109,14 @@ export function waitForChartLoad(chart: ChartSpec) {
       },
     };
   }
+
+  export function resize(selector: string) {
+    return {
+      to(cordX: number, cordY: number) {
+        cy.get(selector)
+          .trigger('mousedown', { which: 1, force: true })
+          .trigger('mousemove', { which: 1, cordX, cordY, force: true })
+          .trigger('mouseup', { which: 1, force: true });
+      },
+    };
+  }
